requirements.txt
requests>=2.32.3
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
praw>=7.8.1
asyncpraw==7.8.1
beautifulsoup4>=4.12.0
gdown>=5.2.0
python-dotenv>=1.0.0
schedule>=1.2.0
feedparser>=6.0.11
transformers>=4.36.0
torch>=2.1.0
flask>=3.0.0
python-telegram-bot>=20.7.0
python-whatsapp-bot>=0.2.0
tweepy>=4.14.0
aiohttp>=3.9.0
asyncio>=3.4.3
python-dateutil>=2.8.2
plotly>=5.18.0
scipy>=1.11.0
threadpoolctl>=3.4.0
sqlalchemy>=2.0.23
